Implement Skija rendered scale widget (#25)
Implementation of the scale widget with custom drawing. Functionality is supposed to be the same as the native scale widgets. The implementation also works on MacOS and Linux. Not implemented is any kind of touch functionality.

The rendering is done by a separate class that is accessed via an interface. See ScaleRenderer and IScaleRenderer. Additional renderer implementations could be easily added. For now ScaleRenderer renders the scale like windows 10/11 would do in light mode.

An isScolled() method was added to Control class as it needs to the Scale widget needs to overwrite the behavior.

---------

Co-authored-by: Dieter Mai <dma@sysgo.com>
Co-authored-by: DieterMaiSysgo <91322115+DieterMaiSysgo@users.noreply.github.com>
Co-authored-by: Heiko Klare <heiko.klare@vector.com>