Revert automatic commit in SkijaGC #60
The current, automated commit mechanism in SkijaGC is insufficient as it
also commits changes in cases where the GC is not used to perform any
drawing operations at all. This leads to unintended effects like
flickering. This change reverts the automatic commit implementation for
now to be replaced by a proper solution later on.

Fixes https://github.com/swt-initiative31/prototype-skija/issues/60