Properly extract size of rendering area for SkijaGC
SkijaGC initialization currently initializes the canvas to draw on with
a size according to the clipping defined for the underlying native GC.
This works in most (if not all) current use cases, but is semantically
incorrect. Once clipping is properly implemented, the clipping change
while using the GC and is only to supposed to define the area to be
considered when doing a paint operation.

This change improves the SkijaGC initialization to initialize a surface
with a size according to the underlying drawable (control, image,
device), which is the maximum area that may be of interest for
rendering.