Windows Canvas: Direct Skija usage
First PoC:
Execute the class window and check the feature.
-------------------------------------------
Explanation
In windows create a memory device context (dc) and point a skija surface
to this dc. Basically it would always be possible to draw to the
surface.
In the paint method, use OS.BitBlt to push the memory area to the
window.
Ideas:
-> It would also be possible to push only specific areas to the window.
So redraw of specific areas is thinkable.
-> Also multiple memory dcs could be used.