Implement proper double-buffering for custom widgets
Currently, the rendering of custom widgets relies on custom implemented
double buffering using an additional image to be painted in. This
requires changes to the widget style (no background) and produces
significant performance overhead.

With this change, native double buffering is used to render custom
widgets. To this end, the SWT.DOUBLE_BUFFERING is properly set for the
controls. The custom double buffering logic is removed from the Drawing
class and previously necessary visibility changes to the style field of
Widgets are reverted. Since the native GC used for double buffering does
not refer to a drawable anymore, it has to be passed to the SkijaGC to
extract bounds information from.