Initialize SkijaGC background from parent instead of with color
Currently, a SkijaGC extracts the background color from the underlying
native GC upon initialization and clears its surface with that color.
This results in backgrounds already rendered into the parent not being
preserved.

With this change, a SkijaGC is initialized with the actual contents of
the already rendered parent by initializing the surface with the
contents from the underlying native GC.