Fix text extent calculation in SkijaGC
SkijaGC text extent calculation did not properly consider trailing
whitespaces. The bounding box calculated by the textExtent operation
does not consider those whitespaces. This change replaces the
calculation to properly consider the actual text width and the height of
the font instead of the height of the actual text to conform with the
behavior of existing GC implementations.